fixed encoding, added field, made differences work

+++ combine_csvs.py
@@ -4,16 +4,25 @@ logger = logging.getLogger(__name__)
 import pandas as pd
 import hashlib
 import datetime
+from dateutil import parser
+
 
 def combine_csvs(csvs_directory,output_directory):
     
     
+    #Keep the 10 most recent files - delete all the others    
     files = os.listdir(output_directory)
     files = [f for f in files if ".csv" in f]
     files = [os.path.join(output_directory,f) for f in files if "__" in f]
     
+
+
+
     for f in files:
-        os.remove(f)
+        basef = os.path.basename(f)
+        delta =  datetime.datetime.now().date() - parser.parse(basef[:8]).date()
+        if delta.days > 2:
+            os.remove(f)
 
     #Get list of files to join
 
@@ -27,12 +36,10 @@ def combine_csvs(csvs_directory,output_directory):
     file_list = [os.path.join(csvs_directory,d) for d in file_list if ".csv" in d]
 
 
-    dt = datetime.datetime.now()
+    
 
-    if dt.time() < datetime.time(12):
-        outfilename = file1[:8] + "__all_current_fhrs_am.csv"
-    else:
-        outfilename = file1[:8] + "__all_current_fhrs_pm.csv"
+    time_formatted = datetime.datetime.now().strftime('%H_%M_%S')
+    outfilename = file1[:8] + "__all_current_fhrs_{}.csv".format(time_formatted)
 
     #Write csv
